python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-tool-router"
version = "0.1.0"
description = "Python wrapper for the MCP tool router daemon"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "MCP Tool Router contributors" }
]
keywords = ["mcp", "tools", "router", "bm25"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3"
]
dependencies = ["pyyaml>=6.0"]

[tool.hatch.build.targets.wheel]
packages = ["mcp_tool_router"]
